requirements.txt
yt-dlp
requests
python-telegram-bot
Flask